requirements.txt
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
pandas-datareader>=0.10.0
gspread>=5.12.0
google-auth>=2.23.0

